+++ mail/mailer.go
@@ -2,6 +2,7 @@ package mail
 
 import (
 	"context"
+	"crypto/rand"
 	"encoding/base64"
 	"fmt"
 	"github.com/cloudogu/ces-importer/configuration"
@@ -45,7 +46,7 @@ type SenderService func(addr string, a smtp.Auth, from string, to []string, msg 
 // Sender provides functionality to send emails using a configured SMTP service.
 type Sender struct {
 	config           configuration.Smtp // SMTP configuration
-	sourceInstance   string             // Source instance URL of the exporter
+	sourceInstance   string             // Source instance URL of the exporters
 	senderService    SenderService      // Function to send email
 	readFile         OsReadFile         // Function to read email content from a file
 	attachments      []string           // List of files to attach to each mail
@@ -82,9 +83,9 @@ func CreateSender(config configuration.Smtp, sourceInstance string, attachments 
 // Returns an error if email composition or sending fails.
 func (s *Sender) Send(ctx context.Context, isFinal bool, migrationResult error, start time.Time, end time.Time) error {
 	slog.Info("Sending migration result via mail...")
-	slog.Info(fmt.Sprintf("Mail is sent from: %s", s.config.From))
-	slog.Info(fmt.Sprintf("Mail is sent to: %v", s.config.To))
-	slog.Info(fmt.Sprintf("Mail is sent to server: %s", s.server()))
+	slog.Debug(fmt.Sprintf("Mail is sent from: %s", s.config.From))
+	slog.Debug(fmt.Sprintf("Mail is sent to: %v", s.config.To))
+	slog.Debug(fmt.Sprintf("Mail is sent to server: %s", s.server()))
 	if s.auth() != nil {
 		slog.Info("Using authentication for mail server")
 	} else {
@@ -100,7 +101,9 @@ func (s *Sender) Send(ctx context.Context, isFinal bool, migrationResult error, 
 	body := s.body(migrationResult, s.sourceInstance, targetInstance, start, end, isFinal)
 	boundary := "MIME_BOUNDARY_CES_IMPORTER"
 	mime := fmt.Sprintf("MIME-Version: 1.0\r\nContent-Type: multipart/mixed; boundary=%s\r\n\r\n", boundary)
-	message := mime +
+	messageId := fmt.Sprintf("Message-ID: %s\r\n", buildMessageId(targetInstance))
+	message := messageId +
+		mime +
 		"--" + boundary + "\r\n" +
 		"Content-Type: text/plain; charset=utf-8\r\n\r\n" +
 		body + "\r\n"
@@ -111,7 +114,7 @@ func (s *Sender) Send(ctx context.Context, isFinal bool, migrationResult error, 
 			slog.Error(fmt.Sprintf("failed to add attachment to mail: %v", err))
 			continue
 		}
-		slog.Info(fmt.Sprintf("Added attachment to mail: %s", attachment))
+		slog.Debug(fmt.Sprintf("Added attachment to mail: %s", attachment))
 		message += attachment
 	}
 
@@ -216,4 +219,16 @@ func chunkSplit(body string, limit int) string {
 		chunked = append(chunked, body[i:end])
 	}
 	return strings.Join(chunked, "\r\n")
+}
+
+func buildMessageId(fqdn string) string {
+	b := make([]byte, 16)
+	_, err := rand.Read(b)
+	if err != nil {
+		slog.Error(fmt.Sprintf("failed to generate random ID: %v", err))
+	}
+
+	timestamp := time.Now().UTC().Format("20060102150405.999999999")
+	identifier := fmt.Sprintf("%s.%x", timestamp, b)
+	return fmt.Sprintf("<%s@%s>", identifier, fqdn)
 }